+++ packages/editor/adf-utils/validator.es.js
@@ -1,0 +1,2 @@
+/* eslint flowtype/require-valid-file-annotation: 0 */
+export * from './dist/es2015/validator';