+++ packages/editor/conversation/examples/3-Mock-Provider.tsx
@@ -1,5 +1,3 @@
-import { ReactionContext } from '@atlaskit/reactions';
-import { MockReactionsClient } from '@atlaskit/reactions/src/client/MockReactionsClient';
 import * as React from 'react';
 import { Demo } from '../example-helpers/DemoPage';
 import { MOCK_USERS } from '../example-helpers/MockData';
@@ -13,12 +11,6 @@ const provider = new ConversationResource({
   user: MOCK_USERS[0],
 });
 
-const reactionClient = new MockReactionsClient();
-
 export default function Example() {
-  return (
-    <ReactionContext client={reactionClient}>
-      <Demo provider={provider} dataProviders={getDataProviderFactory()} />
-    </ReactionContext>
-  );
+  return <Demo provider={provider} dataProviders={getDataProviderFactory()} />;
 }