+++ packages/media/media-picker/src/popup/middleware/__tests__/unit/analyticsProcessing.spec.ts
@@ -7,7 +7,6 @@ import {
 import { Action, Dispatch } from 'redux';
 import { State } from '../../../domain';
 import analyticsProcessing from '../../analyticsProcessing';
-import { source } from '../../analyticsHandlers';
 import { mockStore } from '../../../mocks';
 import { showPopup } from '../../../actions/showPopup';
 import { editorShowImage } from '../../../actions/editorShowImage';
@@ -76,7 +75,6 @@ const makePayloadForOperationalFileUpload = (
     },
     ...attributes,
   },
-  source,
   eventType: OPERATIONAL_EVENT_TYPE,
 });
 
@@ -100,7 +98,6 @@ const makePayloadForTrackFileConversion = (
     ...attributes,
     failReason,
   },
-  source,
   eventType: TRACK_EVENT_TYPE,
 });
 
@@ -409,7 +406,6 @@ describe('analyticsProcessing middleware', () => {
           },
           ...attributes,
         },
-        source,
         eventType: OPERATIONAL_EVENT_TYPE,
       },
     );