+++ packages/search/global-search/src/components/jira/JiraQuickSearchContainer.tsx
@@ -47,7 +47,7 @@ import {
 } from '../../api/CrossProductSearchClient';
 import performanceNow from '../../util/performance-now';
 
-const AdvancedSearchContainer = styled.div`
+const NoResultsAdvancedSearchContainer = styled.div`
   margin-top: ${4 * gridSize()}px;
 `;
 
@@ -124,16 +124,16 @@ export class JiraQuickSearchContainer extends React.Component<
         renderNoRecentActivity={() => (
           <>
             <FormattedHTMLMessage id="global-search.jira.no-recent-activity-body" />
-            <AdvancedSearchContainer>
+            <NoResultsAdvancedSearchContainer>
               <JiraAdvancedSearch
                 query={query}
                 analyticsData={{ resultsCount: 0, wasOnNoResultsScreen: true }}
               />
-            </AdvancedSearchContainer>
+            </NoResultsAdvancedSearchContainer>
           </>
         )}
         renderAdvancedSearchGroup={(analyticsData?) => (
-          <StickyFooter>
+          <StickyFooter style={{ marginTop: `${2 * gridSize()}px` }}>
             <JiraAdvancedSearch
               analyticsData={analyticsData}
               query={query}